Finally found an approach to doctest dependecies thats decent
