+++ .cargo/config.toml
@@ -1,0 +1,2 @@
+[build]
+rustflags = ["--cfg", "steamlocate_doctest"]
