+++ Cargo.toml
@@ -10,6 +10,10 @@ readme = "README.md"
 keywords = ["steam", "vdf", "appmanifest", "directory", "steamapps"]
 categories = ["os", "hardware-support", "filesystem", "accessibility"]
 
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs", "--cfg", "steamlocate_doctest"]
+
 # TODO: test more features in CI
 
 [features]
@@ -24,6 +28,11 @@ serde = { version = "1.0", features = ["derive"] }
 
 crc = { version = "3.0", optional = true }
 
+# Custom cfg used to enable a dependency only needed for doctests
+[target.'cfg(steamlocate_doctest)'.dependencies]
+tempfile = "3.8.1"
+
+# Platform-specific dependencies used for locating the steam dir
 [target.'cfg(target_os="windows")'.dependencies]
 locate_backend = { package = "winreg", version = "0.51", optional = true }
 [target.'cfg(not(target_os="windows"))'.dependencies]
