+++ src/tests/legacy.rs
@@ -1,24 +1,17 @@
 use std::convert::TryInto;
 
-use super::helpers::{SampleApp, TempSteamDir, TestError, TestResult};
-use crate::Error;
+// TODO: steamlocate_tempfile cfg for docs. Otherwise rely on a env var to get passed in
+
+use crate::{
+	Error,
+	tests::{TestError, TestResult, helpers::{SampleApp, TempSteamDir, expect_test_env}},
+};
 
 static GMOD_ID: u32 = SampleApp::GarrysMod.id();
 
-// The legacy test env assumed the following prerequisites:
-// - Steam must be installed
-// - At least two library folders must be setup (the steam dir acts as one)
-// - Garry's Mod along with at least one other steam app must be installed
-pub fn legacy_test_env() -> std::result::Result<TempSteamDir, TestError> {
-    TempSteamDir::builder()
-        .app(SampleApp::GarrysMod.into())
-        .library(SampleApp::GraveyardKeeper.try_into()?)
-        .finish()
-}
-
 #[test]
 fn find_library_folders() -> TestResult {
-    let tmp_steam_dir = legacy_test_env()?;
+    let tmp_steam_dir = expect_test_env();
     let steam_dir = tmp_steam_dir.steam_dir();
     assert!(steam_dir.libraries().unwrap().len() > 1);
     Ok(())
@@ -26,7 +19,7 @@ fn find_library_folders() -> TestResult {
 
 #[test]
 fn find_app() -> TestResult {
-    let tmp_steam_dir = legacy_test_env()?;
+    let tmp_steam_dir = expect_test_env();
     let steam_dir = tmp_steam_dir.steam_dir();
     let steam_app = steam_dir.app(GMOD_ID).unwrap();
     assert_eq!(steam_app.unwrap().app_id, GMOD_ID);
@@ -35,7 +28,7 @@ fn find_app() -> TestResult {
 
 #[test]
 fn app_details() -> TestResult {
-    let tmp_steam_dir = legacy_test_env()?;
+    let tmp_steam_dir = expect_test_env();
     let steam_dir = tmp_steam_dir.steam_dir();
     let steam_app = steam_dir.app(GMOD_ID)?.unwrap();
     assert_eq!(steam_app.name.unwrap(), "Garry's Mod");
@@ -44,7 +37,7 @@ fn app_details() -> TestResult {
 
 #[test]
 fn all_apps() -> TestResult {
-    let tmp_steam_dir = legacy_test_env()?;
+    let tmp_steam_dir = expect_test_env();
     let steam_dir = tmp_steam_dir.steam_dir();
     let mut libraries = steam_dir.libraries().unwrap();
     let all_apps: Vec<_> = libraries
@@ -63,7 +56,7 @@ fn all_apps() -> TestResult {
 
 #[test]
 fn all_apps_get_one() -> TestResult {
-    let tmp_steam_dir = legacy_test_env()?;
+    let tmp_steam_dir = expect_test_env();
     let steam_dir = tmp_steam_dir.steam_dir();
 
     let mut libraries = steam_dir.libraries().unwrap();
