+++ src/tests/mod.rs
@@ -1,3 +1,7 @@
 pub mod helpers;
 #[cfg(test)]
 mod legacy;
+mod temp;
+
+pub type TestError = Box<dyn std::error::Error>;
+pub type TestResult = Result<(), TestError>;
